[Rgen] Start implementing the missing parts of the data model for the transformer. (#22091)
The transformer and the code generation share 90% of the data model, but
we need a 10% that is used to fix the mismatch between two projects.

In this change we finish the implementation fo the TypeInfo struct in
the transformer by providing the TryCreate method that will use the
attribute data to decide if a type was marked as nullable or not.

Remember that in the old xamarin bindings we did not use the ?
annotation but we use the NullAllowedAttribute.