[RGen] Add a new flag to determine if a type is a core image filter. (#22110)
In the new binding API we want to be able to differentiate between a
Class and a CoreImageFilter. This PR adds a new flag that can be used by
the generator to identify the binding type and route the code generation
to the correct emitter. This change allows the API to look as follows:

```csharp

// the following binding represents a class
[BindingType<Class> (Name="MyNSObject")
public partial class MyNSObjectClass : NSObject {
    // export methods, properties, events etc..
}

// the following binding represents a core Image filter
[BindingType<CoreImageFilter> (Name="MyNSObject")
public partial class MyNSObjectClass : NSObject {
    // export methods, properties, events etc..
}
```
The change also allows to customize the visibility of the core image
filters like the following example:
```csharp
[BindingType<CoreImageFilter> (Name="MyNSObject", DefaultCtorVisibility=MethodAttributes.PrivateScope)
public partial class MyNSObjectClass : NSObject {
    // export methods, properties, events etc..
}
```

The analyzer will ensure (in a future PR) that a user cannot customize
the visibility of the core Image filter constructors if the binding is
not using the correct flag.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>