[RGen] Add factory method needed to create a aux variable for a BindFrom(typeof(NSNumber)). (#22111)
This change allows to be able to generate an aux variable that will be
needed if a customer defined a export method in the following way:

```csharp

[Export<Method> ()]
public virtual partial void MyMethod ([BindFrom (typeof (NSNumber))] double myDouble);
```

The above binding API indicates that the native method in objc will take
a NSNumber but we want to expose our users a double. In order to send a
NSNumber to objc, we need to declare an aux variable that will be used
for the native method invocation.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>