[Rgen] Add a new CreateBlock method for several concatenated blocks. (#22115)
There are ocassions in which in we want to be able to concatenate
several blocks one after the other. The most common scenario is when we
are using fixed or using blocks. To make the code cleaner we do not want
to create a collection of nested calls like:
```csharp
using (var m1 = new MemoryStream())
{
  using (var m2 = new MemoryStream())
  {
    using (var m3 = new MemoryStream())
    {
      // this is an example with several usings
    }
  }
}
```
instead we want the generator to write:
```csharp
using (var m1 = new MemoryStream())
using (var m2 = new MemoryStream())
using (var m3 = new MemoryStream())
{
  // this is an example with several usings
}
```

This new CreateBlock method takes an collection of lines and writes a
single block similar to the one in the above example.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>