[Rgen] Add a factory method to create the aux variable for a BindFrom(typeof(NSValue)). (#22112)
This change allows to be able to generate an aux variable that will be
needed if a customer defined a export method in the following way:

```csharp

[Export<Method> ()]
public virtual partial void MyMethod ([BindFrom (typeof (NSValue))] CGPoint point);
```

The above binding API indicates that the native method in objc will take
a NSValue but we want to expose our users a CGPoint. In order to send a
NSValue to objc, we need to declare an aux variable that will be used
for the native method invocation.