[RGen] Add a factory method to create the aux variable for a BindFrom (typeof(NSString)). (#22113)
This change allows to create the aux variable that will allow to write a
binding in the following way:

```csharp

[BindingType]
public enum MyEnum {
   [Field<Enum> ("FIRST)]
   First,
   [Field<Enum> ("SECOND")]
   Second,
}

[Export<Method> ()]
public virtual partial void MyMethod ([BindFrom (typeof (NSString))] MyEnum point);
```

The abode biding declaration does the following:

1. Declares a smart enum with 2 values that have a string backing field.
2. Exposes a objc method as a method that takes the smart enum 'MyEnum'

In order to be able to call the objc native method, we need to create an
auxiliary variable that will contain the NSString value of the enum
field. Smart enums, when generated have an extension method called
'GetValue' that does exactly what we need, it returns the field nsstring
value. This new method does exactly that, it returns the call to the
'GetValue' method from the enum.