[Rgen] Update tabbed string building API.
This is an intermediate step to move to use a IndentedTextWriter
internally. To make things more understandable we are renaming the
Append and AppendLine methods.

We are also moving the class to its own space so that we can provide
more than one implementation. One being the string builder based one and
a second one based on a StreamWriter so that we can provide a class to
be used with the transformer to write async files.