[Rgen] Move the tab string builder to use a IndentedTextWriter internally.
This delegates the tab writting to the IndentedTextWriter, but most
importanly allows have the writting common code in a single place. This
change opens the door to implemented a TabWritter<StreamWriter> which
will be used to perform async io to the file system for the Transformer.

The API is identical and tests pass with minimun changes:

1. Updated the use of ToString to be ToCode.
2. The indent value is an int instead of a uint.