[RGen] Provide async methods for the TabbedWriter. (#22121)
Add all the async methods than can be added to the class. Do remember
the following about async methods during the review:

1. We cannot use 'ref' parameters for a async method.
2. We cannot use ref readonly structs parameters and Span is one.
3. We cannot use those same types inside the body of an asyc method. So
WriteRaw has to do the string allocations.

The Async methods are meant for the transformer, which is a project in
which we can use async IO and we do not need to worry about memory in
the same way we do in the roslyn code generator.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>