[RGen] Keep track in the generator if a type needs a stret call.
This changes a little the way we create parameters and types because we
need the compilation target to know if a type needs a stret call.

We do not use the NeedsStret in the equlas method because that
information has not value when we are deciding if a type needs to be
regenerated or not by the roslyn code generator.