[Rgen] Add the initial SDK parsing to the transformer. (#22155)
Add code that will parse a binding and will create the data model. This
PR only focuses on parsing interfaces and ensuring that the correct
binding type is used and that the attributes are correctly parsed.

We are ignoring methods, constructors, properties and events which will
be added in other PRS to keep diffs easier to review.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>