[RGen] Check that properties in the UI namespaces are in the right thread.
When generating code for classes that are in the UI namespaces we must
add a check to ensure that the propety is executed in the UI thread
unless the class has been marked as thread safe.

Tests were added for both cases. We ensure that the correct thread check
is added per platform when generating in a UI namespace, and we make
sure that it does not get added when the class is marked with the
correct flag.