[Rgen] Fix flacky tests when doing several compilations in parallel. (#22195)
The code generation will not run several compilations in memory at the
same time, but that does happen in the tests.

We have some flacky tests that happen when we have 2 or more compilation
at the same time that have different UI namespaces, for example, a MacOS
and a iOS compilation. In the situation of having more than one
compilation we have a race condition on the extension method that
returns the UI namespaces, this is only a problem during the tests. To
workaround it, we won't allow the code generation tests run in parallel.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>