[Rgen] Add a property invocations struct to be used with the syntax factory. (#22191)
This will allow to have a struct that will contain all the needed
invocation data.

We need to update some of the uses from a CompilationUnit to a
StatementSyntax, which actually makes the code better overall while
letting us use a throw statement as the invocation statement for a
property.