[Rgen] Boolean properties do not need a temp variable.
The use of a temp variable in boolean properties was due to a limitation
of how bgen wrote the code. We copied the behaviour to make sure we were
compatible, yet that was not needed.

This change adds supports to boolean getters in properties without the
need of a temp var.