[BGen] Simplify BindAs for smart enums.
This changes does two things:

1. Adds a GetValue internal helper for non value backend smart enums.
2. Uses the GetValue functions as the builder for the NSArray.

The new method (1) is only added to those smart enums that use a
none value type backing field that looks like the following:

```csharp
internal static ASAuthorizationProviderAuthorizationOperation GetValue (NativeHandle handle)
{
  using var str = Runtime.GetNSObject<NSString> (handle);
  return GetValue (str);
}
```

This is later used to retrieve the enum values from the handle in a
BindAs decorated method/property. This is very similar to what we did in
https://github.com/xamarin/xamarin-macios/pull/22216 and has the same
performance considerations (there is no performance change after C# 11
and later).